Add files via upload
MyList.py contains an implementation of a list class in python. This repo gives developers functional boiler plate code to test different under-the-hood implementations of a list class in python. The file experiments.py tests and documents the efficiency of different methods of appending and resizing an array once it exceed its capacity. The results from my implementations are in the scores.txt file.